more fixes for time functions
