+++ app/utils/timeFunctions.js
@@ -1,20 +1,63 @@
+exports.addMinsToTime = (mins, time) => {
+  let temp = new Date();
+  var [timeHrs, timeMins] = this.getHoursAndMinsFromTime(time);
+  temp.setHours(timeHrs);
+  temp.setMinutes(timeMins + mins);
+  [timeHrs, timeMins] = this.getHoursAndMinsFromTime(temp.toLocaleTimeString());
+  console.log(temp.toLocaleTimeString());
+  console.log(timeHrs);
+  console.log(timeMins);
+
+  // // get the times hour and min value
+  // var [timeHrs, timeMins] = this.getHoursAndMinsFromTime(time);
+
+  // // time arithmetic (addition)
+  // if (timeMins + mins >= 60) {
+  //   var addedHrs = parseInt((timeMins + mins) / 60);
+  //   timeMins = (timeMins + mins) % 60;
+  //   if (timeHrs + addedHrs > 23) {
+  //     timeHrs = (timeHrs + addedHrs) % 24;
+  //   } else {
+  //     timeHrs += addedHrs;
+  //   }
+  // } else {
+  //   timeMins += mins;
+  // }
+
+  // make sure the time slots are padded correctly
+  return (
+    String("00" + timeHrs).slice(-2) +
+    ":" +
+    String("00" + timeMins).slice(-2) +
+    ":00"
+  );
+};
+
 exports.subtractMinsFromTime = (mins, time) => {
-  // get the times hour and min value
-  var [timeHrs, timeMins] = getHoursAndMinsFromTime(time);
+  let temp = new Date();
+  var [timeHrs, timeMins] = this.getHoursAndMinsFromTime(time);
+  temp.setHours(timeHrs);
+  temp.setMinutes(timeMins - mins);
+  [timeHrs, timeMins] = this.getHoursAndMinsFromTime(temp.toLocaleTimeString());
+  console.log(temp.toLocaleTimeString());
+  console.log(timeHrs);
+  console.log(timeMins);
 
-  // time arithmetic (subtraction)
-  if (timeMins - mins <= 0) {
-    var subtractedHrs = parseInt((timeMins - mins) / 60);
-    timeMins = ((timeMins - mins) % 60) + 60;
+  // //TODO test when hour should be -1 but it's -0
 
-    if (timeHrs - subtractedHrs < 0) {
-      timeHrs = ((timeHrs - subtractedHrs) % 24) + 24;
-    } else {
-      timeHrs -= subtractedHrs;
-    }
-  } else {
-    timeMins -= mins;
-  }
+  // // time arithmetic (subtraction)
+  // if (timeMins - mins <= 0) {
+  //   var subtractedHrs = parseInt((timeMins - mins) / 60);
+  //   timeMins = ((timeMins - mins) % 60) + 60;
+
+  //   if (timeHrs - subtractedHrs < 0) {
+  //     timeHrs = ((timeHrs - subtractedHrs) % 24) + 24;
+  //   } else {
+  //     timeHrs -= subtractedHrs;
+  //   }
+  // } else {
+  //   timeMins -= mins;
+  // }
 
   // make sure the time slots are padded correctly
   return (
@@ -51,4 +94,12 @@ exports.toSQLDate = (date1, date2) => {
   first = date1.toISOString().slice(0, 19).replace("T", " ");
   last = date2.toISOString().slice(0, 19).replace("T", " ");
   return { first, last };
+};
+
+exports.formatDate = (date) => {
+  let formattedDate =
+    date.toISOString().substring(5, 10) +
+    "-" +
+    date.toISOString().substring(0, 4);
+  return formattedDate;
 };
